requirements.txt
python-dotenv
langchain-core>=0.3.0
langgraph>=0.2.27
langchain-openai>=0.2.0